Promoted pod_2_watch agent in preparation for promotion of atc agent (elimination of Apple Watch base class)
