+++ lib/user_agent/browsers.rb
@@ -15,7 +15,6 @@ require 'user_agent/browsers/apple_home_pod'
 require 'user_agent/browsers/apple_podcasts'
 require 'user_agent/browsers/apple_watch/base'
 require 'user_agent/browsers/apple_watch/atc'
-require 'user_agent/browsers/apple_watch/pod_2_watch'
 require 'user_agent/browsers/audible'
 require 'user_agent/browsers/audio_clip'
 require 'user_agent/browsers/beyond_pod'
@@ -49,6 +48,7 @@ require 'user_agent/browsers/pandora'
 require 'user_agent/browsers/pandora_rss_crawler'
 require 'user_agent/browsers/playstation'
 require 'user_agent/browsers/pocket_casts'
+require 'user_agent/browsers/pod_2_watch'
 require 'user_agent/browsers/pod_mn'
 require 'user_agent/browsers/podbean'
 require 'user_agent/browsers/podcast_addict'
@@ -135,7 +135,7 @@ class UserAgent
       Libavformat,
       OkHttp,
       AppleWatch::ATC,
-      AppleWatch::Pod2Watch,
+      Pod2Watch,
       ApplePodcasts,
       AndroidDownloadManager,
       AudioClip,
